added inProgress to Stage and isCompleted to TreatmentPlan, change Stage percentComplete to double

+++ src/com/cggcoding/models/Completable.java
@@ -9,5 +9,5 @@ public interface Completable {
 	
 	boolean isCompleted();
 
-	int getPercentComplete();
+	double getPercentComplete();
 }
